Clean up and optimize code structure

+++ Repositories/Implementations/UserRepository.cs
@@ -5,46 +5,47 @@ using Reunite.Repositories.Interfaces;
 
 namespace Reunite.Repositories.Implementations
 {
-    public class UserRepository : IUserRepository
+    public class UserRepository(ReuniteDbContext context) : IUserRepository
     {
 
-        private readonly ReuniteDbContext context;
+        private readonly ReuniteDbContext context = context;
 
-        public UserRepository(ReuniteDbContext context)
+        public async Task<List<ReuniteUser>> GetUsersAsync() => await context.Users.Include(u => u.Queries).ToListAsync();
+
+        public async Task<ReuniteUser?> GetUserAsync(string id) => await context.Users.Include(u => u.Queries).FirstOrDefaultAsync(c => c.Id == id);
+
+        public async Task AddUserAsync(ReuniteUser user)
         {
-            this.context = context;
+            context.Users.Add(user);
+            await SaveChangesAsync();
         }
 
-        public async Task<List<ReuniteUser>> GetUsers()
+        public async Task UpdateUserAsync(string id, ReuniteUser newUser)
         {
-            return await context.Users.Include(u => u.Queries).ToListAsync();
+            var existingUser = await GetUserAsync(id);
+
+            if (existingUser is null)
+                return;
+
+            context.Entry(existingUser).CurrentValues.SetValues(new { newUser.Username, newUser.Email });
+
+            await SaveChangesAsync();
         }
 
-        public async Task<ReuniteUser> GetUser(string id)
+        public async Task<bool> DeleteUserAsync(string id)
         {
-            return await context.Users.Include(u => u.Queries).FirstOrDefaultAsync(c => c.Id == id);
-        }
+            var user = await GetUserAsync(id);
 
-        public async Task AddUser(ReuniteUser user)
-        {
-            await context.Users.AddAsync(user);
-            await context.SaveChangesAsync();
-        }
+            if (user is null)
+                return false;
 
-        public async Task UpdateUser(string id, ReuniteUser newUser)
-        {
-            var user = await GetUser(id);
-            user.Username = newUser.Username;
-            user.Email = newUser.Email;
-            await context.SaveChangesAsync();
-        }
-
-        public async Task DeleteUser(string id)
-        {
-            var user = await GetUser(id);
             context.Users.Remove(user);
-            await context.SaveChangesAsync();
+            await SaveChangesAsync();
+
+            return true;
         }
+
+        public async Task SaveChangesAsync() => await context.SaveChangesAsync();
 
     }
 }
